+++ hugegraph-api/src/main/java/com/baidu/hugegraph/api/graph/BatchAPI.java
@@ -29,9 +29,10 @@ import org.slf4j.Logger;
 import com.baidu.hugegraph.HugeException;
 import com.baidu.hugegraph.HugeGraph;
 import com.baidu.hugegraph.api.API;
-import com.baidu.hugegraph.api.schema.Checkable;
 import com.baidu.hugegraph.config.HugeConfig;
 import com.baidu.hugegraph.config.ServerOptions;
+import com.baidu.hugegraph.define.Checkable;
+import com.baidu.hugegraph.define.UpdateStrategy;
 import com.baidu.hugegraph.metrics.MetricsUtil;
 import com.baidu.hugegraph.server.RestServer;
 import com.baidu.hugegraph.structure.HugeElement;
