+++ hugegraph-api/src/main/java/com/baidu/hugegraph/api/job/GremlinAPI.java
@@ -44,10 +44,10 @@ import com.baidu.hugegraph.HugeException;
 import com.baidu.hugegraph.HugeGraph;
 import com.baidu.hugegraph.api.API;
 import com.baidu.hugegraph.api.filter.StatusFilter.Status;
-import com.baidu.hugegraph.api.schema.Checkable;
 import com.baidu.hugegraph.backend.id.Id;
 import com.baidu.hugegraph.backend.query.Query;
 import com.baidu.hugegraph.core.GraphManager;
+import com.baidu.hugegraph.define.Checkable;
 import com.baidu.hugegraph.job.Job;
 import com.baidu.hugegraph.job.JobBuilder;
 import com.baidu.hugegraph.metrics.MetricsUtil;
