+++ hugegraph-api/src/main/java/com/baidu/hugegraph/define/Checkable.java
@@ -1,4 +1,4 @@
-package com.baidu.hugegraph.api.schema;
+package com.baidu.hugegraph.define;
 
 public interface Checkable {
